refactor: move components to directory
